Revert "migrate from aws -> llamao"
