+++ babel.config.js
@@ -1,0 +1,9 @@
+module.exports = function (api) {
+  api.cache(true);
+
+  return {
+    presets: ['@babel/preset-typescript'],
+    plugins: [['@babel/plugin-transform-runtime']],
+    sourceType: 'unambiguous',
+  };
+};
