+++ scripts/nft_trades_blacklist.js
@@ -1,10 +1,13 @@
-const { pgp, indexa } = require('../src/utils/dbConnection');
+const { pgp, connect } = require('../src/utils/dbConnection');
 const blacklist = require('./blacklist');
 
+const db = 'indexa';
 const schema = 'ethereum';
 const table = 'nft_trades_blacklist';
 
 const insert = async (payload) => {
+  const conn = await connect(db);
+
   const cs = new pgp.helpers.ColumnSet(['transaction_hash'], {
     table: new pgp.helpers.TableName({
       schema,
@@ -16,7 +19,7 @@ const insert = async (payload) => {
     pgp.helpers.insert(payload, cs) +
     ' ON CONFLICT(transaction_hash) DO UPDATE SET transaction_hash = EXCLUDED.transaction_hash';
 
-  const response = await indexa.result(query);
+  const response = await conn.result(query);
 
   if (!response) {
     return new Error(`Couldn't insert into ${schema}.${table}`, 404);
