--- src/api/server.js
@@ -1,5 +1,0 @@
-const app = require('./app');
-const port = process.env.PORT || 3000;
-app.listen(port, () => {
-  console.log(`Server listening on port ${port}`);
-});
