+++ src/controllers/eventLogs.js
@@ -1,7 +1,5 @@
 const minify = require('pg-minify');
 
-const { pgp, indexa } = require('../utils/dbConnection');
-
 const query = `
 SELECT
   encode(e.transaction_hash, 'hex') AS transaction_hash,
@@ -147,192 +145,4 @@ const getEvents = async (task, startBlock, endBlock, config) => {
   return response;
 };
 
-const getMaxBlock = async (task, table) => {
-  const query = minify(
-    `
-SELECT
-    MAX(block_number)
-FROM
-    $<table:raw>
-  `,
-    { compress: true }
-  );
-
-  const response = await task.query(query, { table });
-
-  if (!response) {
-    return new Error('getMaxBlock failed', 404);
-  }
-
-  return response[0].max;
-};
-
-const buildInsertQ = (payload) => {
-  const columns = [
-    'transaction_hash',
-    'log_index',
-    'contract_address',
-    'topic_0',
-    'block_time',
-    'block_number',
-    'exchange_name',
-    'collection',
-    'token_id',
-    'amount',
-    'sale_price',
-    'eth_sale_price',
-    'usd_sale_price',
-    'payment_token',
-    'seller',
-    'buyer',
-    'aggregator_name',
-    'aggregator_address',
-  ];
-
-  const cs = new pgp.helpers.ColumnSet(columns, {
-    // column set requries tablename if schema is not undefined
-    table: new pgp.helpers.TableName({
-      schema: 'ethereum',
-      table: 'nft_trades',
-    }),
-  });
-
-  const query = pgp.helpers.insert(payload, cs);
-
-  return query;
-};
-
-const insertTrades = async (payload) => {
-  const query = buildInsertQ(payload);
-
-  const response = await indexa.result(query);
-
-  if (!response) {
-    return new Error(`Couldn't insert into ethereum.nft_trades`, 404);
-  }
-
-  return response;
-};
-
-// used when refilling (in case of adapter bug, missing events etc)
-// deletes trades in nft_trades for a given marketplace (its address(es), event signatures and block range)
-const buildDeleteQ = () => {
-  const query = `
-  DELETE FROM
-      ethereum.nft_trades
-  WHERE
-      contract_address in ($<contractAddresses:csv>)
-      AND topic_0 in ($<eventSignatureHashes:csv>)
-      AND block_number >= $<startBlock>
-      AND block_number <= $<endBlock>
-  `;
-
-  return query;
-};
-
-// --------- transaction query
-const deleteAndInsertTrades = async (payload, config, startBlock, endBlock) => {
-  // build queries
-  const deleteQuery = buildDeleteQ();
-  const insertQuery = buildInsertQ(payload);
-
-  // required for the delteteQ
-  const eventSignatureHashes = config.events.map(
-    (e) => `\\${e.signatureHash.slice(1)}`
-  );
-  const contractAddresses = config.contracts.map((c) => `\\${c.slice(1)}`);
-
-  return indexa
-    .tx(async (t) => {
-      // sequence of queries:
-      // 1. delete trades
-      const q1 = await t.result(deleteQuery, {
-        contractAddresses,
-        eventSignatureHashes,
-        startBlock,
-        endBlock,
-      });
-
-      // 2. insert trades
-      const q2 = await t.result(insertQuery);
-
-      return [q1, q2];
-    })
-    .then((response) => {
-      // success, COMMIT was executed
-      return {
-        status: 'success',
-        data: response,
-      };
-    })
-    .catch((err) => {
-      // failure, ROLLBACK was executed
-      console.log(err);
-      return new Error('Transaction failed, rolling back', 404);
-    });
-};
-
-const insertWashTrades = async (start, stop) => {
-  const query = minify(`
-INSERT INTO ethereum.nft_wash_trades (transaction_hash, log_index)
-WITH counted_trades AS (
-    SELECT
-        transaction_hash,
-        log_index,
-        buyer,
-        seller,
-        collection,
-        token_id,
-        COUNT(*) OVER (PARTITION BY seller, collection, token_id) AS seller_count,
-        COUNT(*) OVER (PARTITION BY buyer, collection, token_id) AS buyer_count
-    FROM
-        ethereum.nft_trades as trades
-    WHERE
-        block_number BETWEEN $<start>
-        AND $<stop>
-        AND NOT EXISTS (
-          SELECT 1
-          FROM ethereum.nft_wash_trades as fake_trades
-          WHERE fake_trades.transaction_hash = trades.transaction_hash
-            AND fake_trades.log_index = trades.log_index
-        )
-)
-SELECT
-    transaction_hash,
-    log_index
-FROM
-    counted_trades
-WHERE
-    buyer = seller
-    OR EXISTS (
-        SELECT
-            1
-        FROM
-            ethereum.nft_trades t
-        WHERE
-            t.seller = counted_trades.buyer
-            AND t.buyer = counted_trades.seller
-            AND t.collection = counted_trades.collection
-            AND t.token_id = counted_trades.token_id
-            AND t.transaction_hash <> counted_trades.transaction_hash
-    )
-    OR seller_count >= 3
-    OR buyer_count >= 3
-  `);
-
-  const response = await indexa.query(query, { start, stop });
-
-  if (!response) {
-    return new Error(`Couldn't insert into ethereum.nft_wash_trades`, 404);
-  }
-
-  return response;
-};
-
-module.exports = {
-  getEvents,
-  getMaxBlock,
-  insertTrades,
-  deleteAndInsertTrades,
-  insertWashTrades,
-};
+module.exports = getEvents;
