+++ src/handlers/getCollections.js
@@ -1,0 +1,6 @@
+const { getCollections } = require('../controllers/collections');
+
+module.exports.handler = async (event, context) => {
+  context.callbackWaitsForEmptyEventLoop = false;
+  return await getCollections();
+};
