+++ src/handlers/getExchangeStats.js
@@ -1,0 +1,6 @@
+const { getExchangeStats } = require('../controllers/nftTrades');
+
+module.exports.handler = async (event, context) => {
+  context.callbackWaitsForEmptyEventLoop = false;
+  return await getExchangeStats();
+};
