+++ src/handlers/getExchangeVolume.js
@@ -1,0 +1,6 @@
+const { getExchangeVolume } = require('../controllers/nftTrades');
+
+module.exports.handler = async (event, context) => {
+  context.callbackWaitsForEmptyEventLoop = false;
+  return await getExchangeVolume();
+};
