+++ src/handlers/getVolume.js
@@ -1,0 +1,6 @@
+const { getVolume } = require('../controllers/nftTrades');
+
+module.exports.handler = async (event, context) => {
+  context.callbackWaitsForEmptyEventLoop = false;
+  return await getVolume();
+};
