+++ src/handlers/triggerCollections.js
@@ -1,80 +1,18 @@
-const _ = require('lodash');
 const axios = require('axios');
 
-const { pgp, nft } = require('../utils/dbConnection');
+const { insertCollections } = require('../controllers/collections');
 const { convertKeysToSnakeCase } = require('../utils/keyConversion');
 
 const sleep = (ms) => {
   return new Promise((resolve) => setTimeout(resolve, ms));
 };
 
-// multi row insert (update on conflict) query generator
-const buildCollectionQ = (payload) => {
-  const columns = [
-    'collectionId',
-    'name',
-    'slug',
-    'image',
-    'tokenStandard',
-    'totalSupply',
-    'projectUrl',
-    'twitterUsername',
-  ].map((c) => _.snakeCase(c));
-
-  const cs = new pgp.helpers.ColumnSet(columns, { table: 'collection' });
-  const query =
-    pgp.helpers.insert(payload, cs) +
-    ' ON CONFLICT(collection_id) DO UPDATE SET ' +
-    cs.assignColumns({ from: 'EXCLUDED', skip: 'collection_id' });
-
-  return query;
+module.exports.handler = async () => {
+  await main();
 };
 
-// multi row insert query generator
-const buildFloorQ = (payload) => {
-  const columns = [
-    'collectionId',
-    'timestamp',
-    'onSaleCount',
-    'floorPrice',
-    'rank',
-  ].map((c) => _.snakeCase(c));
-
-  const cs = new pgp.helpers.ColumnSet(columns, { table: 'floor' });
-  return pgp.helpers.insert(payload, cs);
-};
-
-// --------- transaction query
-const insertCollections = async (payload) => {
-  // build queries
-  const collectionQ = buildCollectionQ(payload);
-  const floorQ = buildFloorQ(payload);
-
-  return nft
-    .tx(async (t) => {
-      // sequence of queries:
-      // 1. config: insert/update
-      const q1 = await t.result(collectionQ);
-      // 2. floor: insert
-      const q2 = await t.result(floorQ);
-
-      return [q1, q2];
-    })
-    .then((response) => {
-      // success, COMMIT was executed
-      return {
-        status: 'success',
-        data: response,
-      };
-    })
-    .catch((err) => {
-      // failure, ROLLBACK was executed
-      console.log(err);
-      return new Error('Transaction failed, rolling back', 404);
-    });
-};
-
-const job = async () => {
+const main = async () => {
+  console.log('trigger Collections handler...\n');
   const api = 'https://api.reservoir.tools';
 
   const apiKey = {
@@ -166,6 +104,5 @@ const job = async () => {
   console.log('insert collections...');
   const response = await insertCollections(payload);
   console.log(response);
+  console.log('done!');
 };
-
-module.exports = job;
